+++ src/engine/utils/Time.java
@@ -1,8 +1,8 @@
 package engine.utils;
 
 public final class Time {
-    private static double lastTime = 0;
-    private static double deltaTime = 0;
+    private static double lastTime = 0.0;
+    private static double deltaTime = 0.0;
     private final static double TIME_TO_NANO = 1_000_000_000.0;
 
     public static void update() {
@@ -16,11 +16,11 @@ public final class Time {
         deltaTime = 0;
     }
 
-    public static double getDeltaTime() {
-        return deltaTime;
+    public static float getDeltaTime() {
+        return (float)deltaTime;
     }
 
-    public static double getFrameRate() {
-        return deltaTime == 0 ? 0 : 1.0 / deltaTime;
+    public static float getFrameRate() {
+        return (float)(deltaTime == 0 ? 0 : 1.0 / deltaTime);
     }
 }
