+++ src/game/entities/enemies/EnemyBee.java
@@ -17,7 +17,7 @@ public class EnemyBee extends Enemy {
     }
 
     @Override
-    protected void updateAction(double dt) {
+    protected void updateAction(float dt) {
         if (this.state != EnemyState.ATTACKING) {
             return;
         }
