+++ src/game/entities/particles/Particle.java
@@ -4,7 +4,6 @@ import engine.elements.entity.Entity;
 import engine.utils.Position;
 import game.Config;
 import game.Galaga;
-
 import java.awt.Color;
 
 public class Particle extends Entity {
@@ -33,7 +32,7 @@ public class Particle extends Entity {
     }
 
     @Override
-    public void update(double dt) {
+    public void update(float dt) {
 
         this.position.moveTo(this.angle, this.speed * (float) dt);
